requirements.txt
flask
flask-cors
pymongo
flask-bcrypt
flask-jwt-extended
python-dotenv
